+++ src/cryptography/hazmat/primitives/asymmetric/ec.py
@@ -151,6 +151,18 @@ class EllipticCurvePublicKey(object):
         Verifies the signature of the data.
         """
 
+    @classmethod
+    def from_encoded_point(cls, curve, data):
+        utils._check_bytes("data", data)
+        if not isinstance(curve, EllipticCurve):
+            raise TypeError("curve must be an EllipticCurve instance")
+
+        if six.indexbytes(data, 0) not in [0x02, 0x03, 0x04]:
+            raise ValueError("Unsupported elliptic curve point type")
+
+        from cryptography.hazmat.backends.openssl.backend import backend
+        return backend.load_elliptic_curve_public_bytes(curve, data)
+
 
 EllipticCurvePublicKeyWithSerialization = EllipticCurvePublicKey